+++ src/Router.js
@@ -1,9 +1,9 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Gnb from "./components/Gnb";
-import HomePage from "./pages/HomePage";
-import SharedPage from "./pages/SharedPage";
-import FolderPage from "./pages/FolderPage";
-import Footer from "./components/Footer";
+import Gnb from "./components/Gnb/Gnb";
+import HomePage from "./pages/HomePage/HomePage";
+import SharedPage from "./pages/SharedPage/SharedPage";
+import FolderPage from "./pages/FolderPage/FolderPage";
+import Footer from "./components/Footer/Footer";
 
 function Router() {
   return (
