+++ src/components/Header/Header.js
@@ -1,4 +1,4 @@
-import "./styles/Header.css";
+import "./Header.css";
 
 function Header({ folderData }) {
   return (
