+++ src/index.js
@@ -2,8 +2,8 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import Router from "./Router";
 import "./index.css";
-import "./reset.css";
-import "./common.css";
+import "./styles/reset.css";
+import "./styles/common.css";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(<Router />);
